Added test for private methods on interfaces Extras:     - Removed unused test examples     - Reordered invokedynamic tests
