+++ master/src/test/java/com/examples/with/different/packagename/dse/invokedynamic/SingleMethodReference.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Lesser General Public
  * License along with EvoSuite. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.examples.with.different.packagename.dse.invokedynamic.dsc.instrument;
+package com.examples.with.different.packagename.dse.invokedynamic;
 
 /**
  * Tests for Method references implementation (InvokeDynamic)
